UMSE Server/requirements.txt
CherryPy==18.1.2
kaitaistruct==0.8
pycryptodome==3.9.4
